api/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asgiref==3.9.1
attrs==25.3.0
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.7.0
idna==3.10
livekit-api==1.0.5
livekit-protocol==1.0.6
multidict==6.6.4
numpy==2.3.3
propcache==0.3.2
protobuf==6.32.1
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
types-protobuf==6.30.2.20250914
typing_extensions==4.15.0
yarl==1.20.1
